Renaming broadcast to impl_broadcast_reduce
